+++ extensions/irc/src/client.test.ts
@@ -1,6 +1,9 @@
 // Irc tests cover client plugin behavior.
-import net from "node:net";
 import { describe, expect, it } from "vitest";
+import {
+  startLoopbackIrcServer,
+  type LoopbackIrcServer,
+} from "../../../test/helpers/irc-loopback-server.js";
 import { buildFallbackNick, buildIrcNickServCommands, connectIrcClient } from "./client.js";
 
 describe("irc client nickserv", () => {
@@ -78,57 +81,6 @@ describe("irc client fallback nick", () => {
     expect(nick).toMatch(/^a+_\d*$/);
   });
 });
-
-type LoopbackIrcServer = {
-  port: number;
-  lines: string[];
-  close(): Promise<void>;
-};
-
-async function startLoopbackIrcServer(): Promise<LoopbackIrcServer> {
-  const lines: string[] = [];
-  const sockets = new Set<net.Socket>();
-  const server = net.createServer((socket) => {
-    sockets.add(socket);
-    socket.setEncoding("utf8");
-    let buffer = "";
-    socket.on("data", (chunk: string) => {
-      buffer += chunk;
-      let idx = buffer.indexOf("\n");
-      while (idx !== -1) {
-        const line = buffer.slice(0, idx).replace(/\r$/, "");
-        buffer = buffer.slice(idx + 1);
-        idx = buffer.indexOf("\n");
-        lines.push(line);
-        if (line.startsWith("USER ")) {
-          socket.write(":server 001 bot :welcome\r\n");
-        }
-      }
-    });
-    socket.on("close", () => {
-      sockets.delete(socket);
-    });
-  });
-  await new Promise<void>((resolve) => {
-    server.listen(0, "127.0.0.1", resolve);
-  });
-  const address = server.address();
-  if (!address || typeof address === "string") {
-    throw new Error("expected loopback IRC server to bind a TCP port");
-  }
-  return {
-    port: address.port,
-    lines,
-    close: async () => {
-      for (const socket of sockets) {
-        socket.destroy();
-      }
-      await new Promise<void>((resolve) => {
-        server.close(() => resolve());
-      });
-    },
-  };
-}
 
 async function collectPrivmsgBodies(server: LoopbackIrcServer, text: string): Promise<string[]> {
   const client = await connectIrcClient({
